Simpleperf to record non-debuggable apps in userdebug build
We should be able to profile all applications in userdebug/eng devices using
simpleperf, including non-debuggable ones. For devices running a release
build, we will still be able to profile only debuggable apps for now.

To achieve that, we just need to make sure to start/stop simpleperf as root
on userdebug/eng phones, except when using startup profiling, as the PID is
not available until later, so we have no option but to pass the --app flag.
However, that shouldn't be a problem, because startup profiling is only
available for apps deployed to the device by the user, so they should be
debuggable.

Change-Id: I6456d4056f70eb48a076968e18fd53e2f56dbde5
Fixes: 113112384
Test: simpleperf_test.cc