PSQ versions of benchmark tests no longer write perf_data
This change prevents BenchmarkTest from writing a perf_data.json file as
long as the --iterations flag is set to 0 for the run. This will keep
tests which don't record any performance data from writing data that
leads to the creation of Perfgate Benchmarks. This is currently
happening when the PSQ versions of the benchmark tests run in
post-submit.

Bug: b/113177619
Test: manually verified that psq tests don't write perf_data.json
Change-Id: I11bfb5b22f1b58e8ec0851ffff1f8e7cac707511