Sync compile / runtime classpaths via Constraints.
Instead of having a task that checks whether the two
classpaths are in sync, use a new Gradle mechanism to
force resolution of compileClasspath based on the
resolution results of runtimeClasspath.

This is enbled by default but can be disabled with
android.dependency.useConstraints=false

Next step is to to the same with prod and test classpaths.

Bug: 113616775
Test: existing tests

Change-Id: I89b04e17406094e26a02bace2a4982f4fd5446e5