Move the packageName from Artifact to Config
This change eliminates redundancy in the model and makes it easier
to autogenerate Artifacts. The various Artifact instances representing
different variants of the same artifact will always have the same
package name.

By moving the package name to the Config, the normal merging of
Configs will cause Artifacts to end up with correct package names
after a call to ConfigTable.generateArtifacts, which removes some
boilerplate from code that instantiates Artifacts.

Test: Existing
Change-Id: I10e67c6af448029cfb932a8f0a7d75428ceb5a6d