Incap: Make data binding incremental (1/3).
Data binding V2 meets all the requirements of an "aggregating"
incremental annotation processor (see
https://docs.gradle.org/5.1.1/userguide/java_plugin.html#sec:incremental_annotation_processing).

For safety, we put this feature behind a flag called
   android.databinding.incremental
to let the users try it. The flag currently defaults to false.

Note that:
  - This feature can only be enabled for data binding V2. It is disabled
    for data binding V1 (and V1 is going to be removed anyway---see bug
    121323067).
  - Data binding is currently an *aggregating* processor. Making data
    binding an *isolating* processor will be even better for build
    performance and is left for future work.

More context:
-------------

The keys to making data binding an aggregating processor are:
   (1) For source code, reading and writing should be done only via the
       JDK annotation processing API.
   (2) For resources (any non-source-code inputs and outputs), they need
       to be registered with Gradle via the CommandLineArgumentProvider
       API---see DataBindingCompilerArguments.kt).

We can confirm (1) by manually inspecting the data binding source code.

(2) was already done a while ago when we made data binding work with the
build cache (bug 69243050).

This CL and others in the same topic add the remaining (housekeeping)
details to make data binding incremental.

Test: New DataBindingIncrementalityTest
Bug: 110061530
Bug: 117725060

Change-Id: I19c0229e9af8fea3ba7719e4e20ceca46bf7268e