Use the signing config from the base module declaration.
Dynamic-features no longer use the signing config from their gradle
file. Instead, it is pulled from the base application since these need
to have the same signing config to deploy onto the device.

The next change will make it a sync error for the signingConfig to be
declared for features.

bug:114626324
Test: Updated unit tests
Change-Id: Ieac8c865e3548055da3417146835ee36a79b9373