Fix failed IncrementalJavaCompileWithAPsTest in Gradle 5.0.
When run with Gradle 5.0, the test fails with two issues:
  1. JavaCompile is expected to be incremental but is not.
  2. Annotation processor output directories are reported as ADDED even
     though they existed in the previous build.

The cause seems to be that we are adding the annotation processor
generated sources to the AndroidJavaCompile task's inputs slightly
differently than the other sources. It seems that Gradle 4.10.x and
earlier allow this but Gradle 5.0 does not.

This CL now adds all sources to the same task input to fix both issues
above.

Bug: 115774858
Test: IncrementalJavaCompileWithAPsTest should pass with Gradle 5.0
Change-Id: I0da8e54234c39d01727740a287b39e3fcffbb179