Expose GoogleMavenRepository collections.
This CL makes the GoogleMavenRepository adds public methods to get lists
of groups and artifacts making it possible to enumerate contents in the
UI. The methods are not supposed ot be called on the UI thread.

Also, the CL fixes a bug in master index parsing with the document root
element being processed as a groupId.

Bug: 113333966
Bug: 116840603
Bug: 116627594
Test: unit test updated
Change-Id: I1452efc2733912970157ddee0768a77357ee9627