IRR: Don't throw exception on remoteEvent
Assimilating remote events means incorporating the remote timeline
into the local timeline. This is impossible to do if the total
duration of remote events is longer than the duration of the local
event.

While it is an unexpected condition, tracing is a debugging feature
which should never ever impact our developers. Instead of wrapping
it in a try/catch, the code now simply log the error.

Bug: None
Test: Manual
Change-Id: I405acee530a020d628af74bf821b84106582f95c