Add support for lib-desugaring and build-system configuration
First, this CL adds explicit options to the lint integration API
for specifying the types of desugaring operations that are
performed. This lets for example the Bazel integration specify
that it is using android_java8_libs.

Second, it adds support for library desugaring, as described in
https://github.com/bazelbuild/bazel/blob/master/tools/android/desugar.sh
This stops the API check from giving false positives for APIs
that are rewritten at build time anyway.

Fixes: 116794191
Test: Unit tests included
Change-Id: I7cf33b86debaf710c5c4499a9b6a6e4d75979426