Add mean tolerance params to default analyzer
Previously, we only had median tolerance params, which allows the
analyzer to be resilient to huge spikes in the data, but which can yield
a lot of false positives for metrics which have bimodal behavior (i.e.,
times tend to be either high or low, not in between)

This CL adds mean tolerance params, which should not yield as many false
positives for metrics with bimodal behavior, since the mean is more stable
than the median for samples from bimodal distributions.

Now the median *and* mean tolerances will have to be exceeded for the analyzer
to flag a regression.

Test: existing
Bug: N/A
Change-Id: Ib5ac3da3379ae0764150c9a419cec8f716f94203