IRR: Roll out more robust class compare system
Problem:
========

So far we relied on the assumption that dex filenames were stable.
Namely, that a class was always located in the same .dex apk entry
. However there have been cases where we receive:

     OLD FILENAME    CRC  |      NEW FILENAME     CRC
     classes1.dex     A   |      classes1.dex      A
     classes2.dex     B   |      classes2.dex      D
     classes3.dex     C   |      classes3.dex      B

classes2.dex and classes3.dex seem to have changed. However
classes2.dex as only been renamed classes3.dex and the only real
changes are between the pair old classes3.dex and new classes2.dex.

Solution:
=========

This problem is fixed by disregarding filename completely. The list
of all classes is flatten across all .dex files and the delta is
built from comparing this flattened list and the content of each
new .dex entry in the new apks.

Bug: 118889251
Bug: 118451016
Test: Manual
Change-Id: Ibba6cba7eae9e3e93e16aad18d1497d00fe23379